moved enemie class
